Add writable module
