+++ node/_stream.js
@@ -4,6 +4,7 @@ import { Stream } from "./internal/streams/legacy.js";
 import { isUint8Array } from "./_util/_util_types.ts";
 import { Buffer } from "./buffer.ts";
 import Readable from "./internal/streams/readable.js";
+import Writable from "./internal/streams/writable.js";
 
 // const {
 //   promisify: { custom: customPromisify },
@@ -17,7 +18,7 @@ import Readable from "./internal/streams/readable.js";
 
 // Stream.isDisturbed = require("internal/streams/utils").isDisturbed;
 Stream.Readable = Readable;
-// Stream.Writable = require("internal/streams/writable");
+Stream.Writable = Writable;
 // Stream.Duplex = require("internal/streams/duplex");
 // Stream.Transform = require("internal/streams/transform");
 // Stream.PassThrough = require("internal/streams/passthrough");
@@ -63,4 +64,10 @@ Stream._isUint8Array = isUint8Array;
 Stream._uint8ArrayToBuffer = _uint8ArrayToBuffer;
 
 export default Stream;
-export { _uint8ArrayToBuffer, isUint8Array as _isUint8Array, Readable, Stream };
+export {
+  _uint8ArrayToBuffer,
+  isUint8Array as _isUint8Array,
+  Readable,
+  Stream,
+  Writable,
+};
